edit cover and desc

+++ back-end/src/routes/games.js
@@ -82,11 +82,7 @@ router.post("/create", tokenAuth, upload.single("coverImage"), async (req, res, 
 );
 
 // Update community
-router.put(
-  "/update/:id",
-  tokenAuth,
-  upload.single("newCover"),
-  async (req, res, next) => {
+router.put( "/update/:id", tokenAuth, upload.single("newCover"), async (req, res, next) => {
     try {
       const id = +req.params.id;
       const isAdmin = req.isAdmin;
@@ -95,10 +91,11 @@ router.put(
           error: "no auth",
         });
       }
-      const { gameName, description, isActive } = req.body;
       const existing = await prisma.gameCommunity.findUnique({ where: { id } });
+      
+      const coverImage = req.file ? `/images/games/${req.file.filename}` : undefined;
 
-      const coverImage = `/images/games/${req.file.filename}`;
+      const { gameName, description, isActive } = req.body;
 
       if (!existing)
         return res.status(404).json({ error: "Community not found" });
@@ -119,11 +116,25 @@ router.put(
 router.delete("/delete/:id", tokenAuth, async (req, res, next) => {
   try {
     const id = +req.params.id;
+    const isAdmin = req.isAdmin
+
+    if(!isAdmin){
+      return res.status(403).json({
+        error: "no priv"
+      })
+    }
+
     const existing = await prisma.gameCommunity.findUnique({ where: { id } });
 
+    console.log("existing => ", existing)
     if (!existing)
       return res.status(404).json({ error: "Community not found" });
 
+    await prisma.post.deleteMany({
+      where: {
+        communityId: id
+      }
+    })
     await prisma.gameCommunity.delete({ where: { id } });
 
     res.status(200).json({ message: "Community deleted" });
